requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
executing==2.1.0
inflection==0.5.1
ipython==8.29.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
wcwidth==0.2.13
